pdm: make stock browser a report dialog
Having this as a dialog will make it possible for user to supply some
parameters.
In the old UI, this was relying on domain tree to let the user select
the sites they want to see, but in ERP5JS domain trees only work on
list mode listboxes.
Also, we have plans to make this browser more interactive, to let the
user choose more parameters (section category, quantity unit for
conversions etc) in the report dialog, but this was not possible in a
view action.