pdm_renderjs_ui_test: align test with pdm_ui_test
Previously, these tests were displaying the listbox with two lines, one
for source one for destination and where checking both lines at once.
Now that it's possible to filter by node category (site) in renderjs as
well, adjust the test to select the node categories one by one and check
that the results for each line category is what we expect.

Also fix a few typos