pdm: allow to select node category in stock browser dialog
This allows to filter the node by site in new and old interface in
a consistent way.

The domain tree is also removed for consistency (so that we don't have
different features on new and old interface)

Old UI tests have also been updated to use more meaningful macro names.