erp5_officejs_support_request_ui: New gadget to generate RSS Link and remove it from dashboard
Having a big "Generate RSS" button on the front page is strange, because the front page button are for common use cases that users are supposed to execute a lot.

For example, makes sense to have a "Submit new Support Request" button, because users are using this app to submit support requests, but generating RSS is very exceptional action, so we wanted to remove it from the front page.