sql_browser: backup editor content in a textarea
Browsers (at least chrome) saves the textarea content and restore it
when going back or resurecting a closed tab, which prevents loosing
unsaved work.
This browser feature does not work in "rich" javascript
editors, but this trick of syncronizing the rich editor content to a
hidden textarea seem to makes it work.