xhtml_style: Use an image for "Sign in with Google" button
According to Google's branding guidelines [1], using zocial like we did is not
OK because the logo is different.

To comply with Google's branding guidelines, we simply use an image from
"Download files" section of the Sign-In Branding Guidelines page. Using a png
image is not really good, but a vector version would require to download a
font, so the image way is easier.

1: https://developers.google.com/identity/branding-guidelines