pdm: use different actions categories for ERP5JS and xhtml
In ERP5JS we want the "interactive reports" to be available like other view
actions, directly in the left side panel. Also "Report" are named "Export" in
this interface (which might actually be a mistake, but it's like this now).
Rendering a form_dialog in a view action works perfectly in ERP5JS, so we can
take advantage of this.

In xhtml_style, this can not be a object_view action, because object_view
actions can not use form_dialog, this breaks the navigation.

So split in two actions, to have the best of both worlds