erp5_officejs_support_request_ui: Add new view action category to avoid conflicts with object_view
The view of support requests in the app is a new "Discussable" page, where users can post comments and attach documents to their posts. If
they attach document, a link to that document is displayed in the discussion thread. The problem was that this link was never really
correct, depending on the portal type of the attached document. This was never really working, because the code here was taking the 1st action
for image and 4th action for documents we believe that this was made to use the "preview" action of the portal type, but depending on the order
of actions we had different results. This was simplified by introducing an action category dedicated for the support request application
(object_officejs_support_request_view), configuring this web site to use this action category and adding on all portal types used in this app the
actions. This way, for all document portal types we have a preview action (and the links from the discussion page directly show the preview
action).