tests: make each SavedTestSuite instance use a different logfile
When using SavedTestSuite, the first step is to make testnode run --save for
each instance, by running the test referenced in _saved_test_id class
attributes.

Because log files are defined only by the test name, this was causing the
multiple concurrent runUnitTest executions with --save to all log in the same
file. That file was also reused when actually running this test later.

Override getLogDirectoryPath to use a different folder, one per instance, when
doing the initial --save