credential: prevent multiple failures
If a Credential Request/Recovery can not be accepted, the alarm should not
retry again and again.

We use an hasActivity test before trying to accept to prevent causing multiple
failing activities (if it already failed once, no need to retry) and maybe some
race conditions if alarm is too slow to run.