OAuthTool: Used only as a Folder ({Google,Facebook}Connector) so remove {unused,unmaintained,unfinished} code.
Properties are statically sets on portal_oauth directly with classes define in
OAuthTool.py. This prevents migration to ZODB Components as migrating OAuthTool.py
to erp5_oauth would mean that upgrading it would fail (chicken & egg issue: we need
these classes importFile(portal_oauth.xml) on portal_templates.download() but they
are not available until the bt5 is installed...).

Considering that several properties are supposed to be private but statically set
in OAuthTool.py, and that this code has never been used since its introduction 9
years ago, better remove it than trying to fix it properly.