pdm: support defining price in another quantity unit
So far price were always assumed to be defined in the default
quantity unit of the resource, even though supply lines had a
quantity unit field, it was ignored. But it can be useful to
define prices using another quantity unit than the default quantity
unit of the resource, for example when a resource is purchased
in kg but managed in m³.

To support such scenario, we adjust the priced quantity unit
when applying price.