ERP5OOo: Migrate FormPrintout and OOoTemplate to Portal Type as classes.
Likewise ERP5Form, there is no reason why these are not ERP5 object yet. This
change is done in preparation of migration to ZODB Components (the Document
will be migrated to ZODB Components thus requiring a dynamic class so that
the XML of FormPrintout instance can be imported).

This also adds the necessary Property Sheets for FormPrintout and OOoTemplate
as this is required by constructInstance() (IOW, properties are not set through
__init__() anymore but through edit()).