test_result: Retry failed test result lines
If a test result finish and its title match the test suite pattern allowed to
restart defined on test suite, it will be retried at most once.

Note about the test: The current behavior is to use PASS/FAIL for test result
and PASSED/FAILED for test result line. The test checks the actual values, if
we ever fix this inconsistency, we can safely update the test I believe.