configurator: Fix some minor errors in BusinessProcessConfiguratorItem
This code was iterating in a dict and removing keys at the same time.

The second block was the same logic as the first, but using another variable
all along, except for the part where None keys are removed from dict. Change
this part to use the other variable used everywhere else in the second block.