erp5_officejs_support_request_ui: Add begin date to the query when user click on "Last Month Activity"
The filter when the user click on the "Last Month Activity" chart filter Support Request only by status (Submitted / Open / Suspended / Closed). So on the table "Recent updates", all the support request with the chosen status appear. There is no filter by "Begin Date" for the last 30 days, which is no coherent with the "Last Month Activity". => It should appear only the support request with the chosen status with a begin date < D - 30 days