ui_test_core: fix verifyImageMatchSnapshot for pages using css fonts
We are using html2canvas to render the page in a canvas element, but
because the canvas element was created in the selenium runner frame (and
not the tested frame), if some extra css fonts were loaded in the custom
page they did not apply.

To make the snapshot use proper fonts, we create the canvas in the
tested page.