BusinessTemplate: update tool if they have properties
Since 248f59e539 (Business Template: Likewise ERP5Site.addERP5Tool(), do
not re-create Tool if it already exists., 2020-06-22) we no longer
update tools, because of the problem that business template does not
currently handle updating objects with lots of sub-objects.

But we realized that we really need to update tools when they contain
configuration as object attributes, like mimetypes_registry, where the
problem was observed.

Instead of unconditionnally skipping any tool during update, we inspect
the tool we are about to install and only skip the tool if it looks
already OK: the __class__ is already correct and the new tool does not
have any properties in their __dict__