BusinessTemplate: update tools, except the one we know to be too big
Since 248f59e539 (Business Template: Likewise ERP5Site.addERP5Tool(), do
not re-create Tool if it already exists., 2020-06-22) we no longer
update tools, because of the problem that business template does not
currently handle updating objects with lots of sub-objects.

But we realized that we really need to update tools when they contain
configuration as object attributes, like mimetypes_registry where the
problem was observed.

Instead of unconditionnally skipping any tool during update, we skip
only portal_simulation, because we know that it will contain lots of
applied rules and portal_preferences, because we know it contain
user preferences.

Other candidates were:
 - portal_trash/portal_templates: contain lots of documents,
proportional to the number of business template updates, but is not
managed in business template (for bootstrap issues)
 - portal_activities: contain a lot of active processes, but is not
managed in business template