base: rename tests for person property acqusition
In 93e30e5ea9 (Person: Store user id in new user_id property., 2016-12-09)
we adapted this test to the new behavior: title fallback to user_id or
id, but since persons always have a user_id by default, this test was
changed to check that title fallbacks to suer id, but the name of the
test still mention "fallback to id" which became a bit different from
what was tested here.

Revert testEmptyTitleFallbackOnId to check that title fallbacks on id,
using persons without user id and introduce new
testEmptyTitleFallbackOnUserId to describe the new behaviour with user
id.