base: ensure migrated persons have a user_id
Persons created before the introduction of ERP5 Login and user_id will only
have a user_id after migration if they were already user before migration,
otherwise they will not have a user_id and creating assignments and ERP5 Login
for this person creates a user which can not log in the system.

To make it possible for these persons to login anyway, we ensure person has a
user id when validating a login

This is bc67c2c4ce8612c27977d79509b110bd20b3edbe again, but with a proxy role
on the interaction workflow and a non regression test to cover the problem with
the first version of that commit.