Zelenium: update key event dispatching to support modern browsers
Use the new KeyboardEvent instead of deprecated KeyEvent and
initKeyEvent, see
https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/initKeyEvent#specifications
and
https://lab.nexedi.com/nexedi/slapos/merge_requests/800#note_115488

In old firefox dispatching a fake keydown event with keycode 13 (enter)
on an input field was enough to submit the corresponding form, like in
a real browser and like in webdriver, but in modern browsers this no
longer submits the form. Adjust the logic to submit the form to mimic
real browser behavior.