corporate_identity_test: adjust test for new "Export as Chapter" action
Since 206a8e250b (erp5_officejs: new OfficeJS Slideshow Editor application,
2020-07-06) this test runs with erp5_run_my_doc installed, and the first export
action for Web Page is no longer "Export as SlideShow" but "Export as Chapter".
This is an action which is here since a long time in erp5_run_my_doc, but since
erp5_run_my_doc was not installed in this test it was not impacting this test.

Adjust the tests so that they don't assume that the first export action is
"Export as SlideShow" but are more independent on which is the first action.

Also, all actions have a new index, so places where we select the action by
index no longer work. Instead of updating the indexes, select by label which
should be more future proof.