forge: fix random failures with testTemplateTool
getRepositoryList contain all business template repositories, ie:

      ['/srv/slapgrid/slappart4/srv/slapos/soft/c63ba7265399450b28f9ea6d5667a5e7/parts/erp5-doc/bt5',
       '/srv/slapgrid/slappart4/srv/slapos/soft/c63ba7265399450b28f9ea6d5667a5e7/parts/erp5/bt5',
       '/srv/slapgrid/slappart4/srv/slapos/soft/c63ba7265399450b28f9ea6d5667a5e7/parts/erp5/product/ERP5/bootstrap',
       '/srv/slapgrid/slappart4/srv/slapos/soft/c63ba7265399450b28f9ea6d5667a5e7/parts/erp5-bin/bt5']

it seems in random order.

Test was written to use the one from parts/erp5/bt5, but since it just
takes "anything which does not contain bootstrap" it was sometimes getting
erp5-bin or erp5-doc and could not find erp5_configurator_standard business
template.

Change the selection of the repository to take into account that we may also
have erp5-doc and erp5-bin and so that it fails if more than one.