xhtml_style: fix invisible cursor in textarea with chrome
With chrome ( 85.0.4183.108 but I think it appears a few weeks/months ago )
when using erp5_xhtml_style when the cursor in the first column on the left
it's not visible, because the bold background hides it.

Default user agent spreadsheet is a 2px padding, but we have a rule applying a
padding 0 for all textarea. Revert this rule for textarea used as fields in
content.