test_result: fix "Retry Count" being acquired on test result line
At the end of test result, we calculate the sum of test result line
retries and set it as a property on test result. Listbox will then
show for each test result line the number of retries. Problem is
that we did not set retry count property on test result lines which
were not retried, so they acquired this from the parent test result.

To prevent this, always set a retry count property on test result
and set it to 0 when they are not retried.