testERP5Catalog: test new related key syntaxes
In 4bcf5c6a32 (CatalogTool: Introduce new dynamic related key syntax, 2017-07-25)
we introduced a new syntax to solve some ambiguities in the previous syntax, but
the test suite was still testing the old syntax only.

Update tests to test the new syntax and to also check the old legacy syntax.