content_translation: Also index translations from message catalogs
Now that we are able to search using content translation, we also want
to extend this so that columns like "group__title" can be searchable in listboxs,
but categories are usually translated with `erp5_content`, which was not indexed
in content_translation table.

Extend content translation indexing to also support translations coming from
Localizer catalog make this possible.

The only limitation I could think of is that categories will need to be reindexed
when the messages are modified in catalog.