web_renderjs_ui: increase space between listbox column header and sort icon
The space between listbox column header and icon used to display the sort order
was a bit too small, making it hard to see if the sort was ascending or descending
sometimes.

To have more space, set a margin-right on the before pseudo-element, like we are
already doing for buttons with text and icon.