core: support Manager user in Base_getDialogSectionCategoryItemList
Base_getDialogSectionCategoryItemList was supporting the case of
users from zodb_users plugin, but not the case of ERP5 users with
roles set from zodb_roles.

It's a more and more common case to make manager users using a
"normal" ERP5 user as person (so that we can reset password etc)
and to give Manager role by zodb_roles, but this was not supported
by this script.