core: fix error when submitting "View Diff" dialog
This action was working fine the first time displayed, but clicking
again on "View Diff" button was causing an IndexError: list index out of range
when used on one document.

This was because in this case list_selection_name exists in REQUEST, but
is an empty string. Instead of checking that this selection name is not
None to detect if we are in module mode or in document mode, check that
its true-ish to also tolerate an empty string.