testnode: advertise log URL with log_frontend_url
testnode use to publish the URL of logs so that Nexedi ERP5 can
display link to "view logs". This URL was an IPv6 URL on a certificate
which does not validate.

Now testnode software release request a shared frontend for these
logs and generate a config file with log_frontend_url set.

testnode now advertise the URL of log using the log frontend url
from config.

Also simplify frontend_url key to manage it using the same way.