ERP5TypeFunctionalTestCase: wait a bit longer for Xvfb to start
When running on testnode, runUniTest is responsible for picking a free $DISPLAY.
This is implemented by starting Xvfb on a $DISPLAY, waiting a bit and if it is
still running we assume the $DISPLAY was free. But it seems sometimes we don't
wait enough, when testnodes are overloaded.

This implementation is wrong, we could run a command check that X is running,
but as explained in the code command below, xdpyinfo or other tools are
currently not available in test environment. Since we are considering switching
this to seleniumserver (which would take care of running a X Server for us) so
this is OK for now.