dms: support password protected PDFs in viewer
This is intended to store sensitive documents that users will only
be able to view from ERP5, but not to easily print or download.

To use this, some customization is needed. First, PDF needs to be
saved with a password. Then, PDF_getContentPassword type based method
needs to be customized to return the password.

When using this, we use different password for each document, by
deriving a password from a master key using document properties (such
as document reference for example)