fixup! Get rid of Products.CMFDefault.{File.File,Portal.CMFSite}
See commit 12e4cc8b2a3fabd4c56fba84afb5b8daded0af55.

File._setFile does not support that 'data' is a bytes object:
it crashes if there are already data stored on the object.
test_PUT_on_web_page (erp5_dms:testWebDavSupport) is changed to cover
this case and the caller of File._edit is fixed to make that 'file'
is always a file-like object.

File._setFile is also optimized to avoid reading data for nothing
if the document has no data.