accounting: Fix timezone issues when checking periods are consecutive
Opening an accounting period is refused if the start date of the period
is more than one day after the stop date of the previous period, but this
check did not take into account that "next day" might be more than 24
hours, like it's the case when daylight saving happen between these dates.
Instead we check that the difference is less than 1.9 days.

Reorganise tests to group accounting period related tests in a dedicated
test class and add missing tests for period validation checks.

Also fix a few race conditions with catalog indexing that are probably not
a problem in real life but were revealed by the test.