smart_assistant: use relative URL for gadget URL
Using an absolute URL for gadgets cause the gadget to have a different URL
per language when the web site uses "Language in URL" feature.

Using absolute URL is not good for cache and also it seems to cause
incompatibilities with recent versions of ERP5JS. Using relative URL is
what we usually do, it does not have cache problem and works fine.