ingestion: make "group" of attached documents easily configurable
Base_viewNewFileDialog/your_publication_section allows to easily suggest
a default publication section to the user, but the "group" was not so
easily customizable, it was just getting from document or user.

Introduce the same mechanism as for publication section, a type based
script that can be defined for each portal type.