web_renderjs_ui: don't use manage_FTPget to get page code
manage_FTPget is an API to serve the file content to response, it is not
suitable to just get the file/page template content like we need here.
It was causing issues when running through activities, because timerserver
responses does not have the setBase method, but is also wrong because
large files are streamed to response directly.

Zope does not really offer an unified API to get the source of page
templates or files, except the PrincipiaSearchSource method which it uses
for its builtin search engine. It seems suitable to use here as well.

Also add a test for the case where web pages are in a OFS.File, only
the case of PageTemplate was covered.