document_scanner: introduce a customizable type based script to suggest a title for scanned document
Unlike when we attach a document, where the title comes from the filename of the attached document,
when we scan document we don't have any filename. There's a default behaviour of using the current date, so that title is not empty, but in some cases we can define a script to suggest a better title, for exemple when scanning an invoice, we can suggest using supplier name and the invoice title.