l10n_fr: fix duplicate "Exporter" action name
On documents, we have an "Export" action for OOoDocument_viewConvertDialog.
When erp5_ods_style is installed, "Export to Spreadsheet" is added as a global
action. Both these actions were translated as "Exporter" in french, which is
confusing and a bit incorrect.
Use a proper, different translation for "Export to Spreadsheet"