accounting: fix grouping option of GL when running in deferred mode
This omit_grouping_reference key was not set as selection parameters, so
Node_getAccountingTransactionList could not find it in selection when running
in deferred mode.

In non deferred mode, it was working, because selection is set with values from
request and it was same request, but deferred style uses activities, so it's
different requests.