base: make BankAccount.getReference use BankAccount.getIban
When IBAN is defined, use it by default for reference. IBAN is used by
default all over europe, so when it's set it should be used by default for
reference.
Mixing IBAN and other properties does not make sense, so this change
should not have unexpected side effects.

We also update "edit_order" on BankAccount_view so that title properly
fallbacks on reference when reference is iban.