deferred_style: Properly keep request parameters
When rendering reports, deferred style was only keeping request.other, but
some places, like Products.ERP5Form.ListBox.ListBox.getParamDict read only in
request.form, which caused different behaviours when rendering reports
directly or in activities, because when rendering in activities request keys
that the report dialog might have pushed were not preset in request. They
were of course present when rendering the report directly, because this was
the same request, so this lead to different behaviours wether reports were
rendering in activities or directly, like we could observe in 7058ab6f88
(accounting: fix grouping option of GL when running in deferred mode, 2021-02-25)

This change deferred style to properly propagate request.other and
request.form in all activities of reports.