erp5_dms: Add modificaiton_date to URL to make sure browser reloads the image in ERP5JS
In ERP5JS, when we access a Web Illustration, change it through Edit tab and go back to View, the image is old because the URL and browser does not ask to backend if something has changed.

For example, changing and switching to View tab never changes the URL "web_page_module/9?format=".

If we access it directly it returns 200 for the first time, 304 if we reload and 200 again if we edit this document. This is good and expected.

To fix this issue, modification_date will be passed in the URL and every change will make this property different. Then, browser will reload the page only if we change it.