corporate_identity_test: use PNG files for reference data
BMP are not compressed and use more disk space than PNG.
The trick is that ERP5 uses by default a quality of 75% on images ([this is default on propertysheet])
so we need to dump the image with maximum (100%) quality.

Update also assertion message to not include the full base64
image in case of difference, use a short assertion message
and save images in log folder.

[this is default on propertysheet]: https://lab.nexedi.com/nexedi/erp5/blob/2f9b67fae3ac59955231d4191a34134ff1b6079a/product/ERP5/bootstrap/erp5_property_sheets/PropertySheetTemplateItem/portal_property_sheets/HtmlStylePreference/preferred_image_quality_property.xml#L56