base: Configure substitution's "ignore missing" on notification messages
TextDocument substitution API already supported an `ignore_missing` argument,
so that we can programmatically control wether missing entries should be
rendered as ${variable} or should raise an error. This makes sense for some
"important" notifications, where not sending a message would be less problematic
than sending a message where some variables have not been substituted.

This extends this concept by allowing to configure as a property on the
notification message wether missing entries should be ignored. This default to
"ignore" to maintain compatbility.