payment_mean: rework "limit" on select payment dialog
Rename the field, because on xhtml_style, submitting the field with an
empty limit now cause an error like:

    ValueError: invalid literal for int() with base 10: ''

and on ERP5JS it is ignoed.

Also, make the field required and set a small default value and a range,
to prevent accidentally displaying/selecting too much when there's lot of
data matching.