erp5_web_renderjs_ui_test: Improve test to expect new notification message
When there is a invalid field, we trigger a notification if the user tries to save the form. But, if this notification is triggered, we don't trigger an alert when the user clicks in another page to leave.

But, if the field is invalid, the user does not click in "Save" and try to leave, the alarm is triggered.